requirements.txt
altair
pandas
streamlit
numpy
matplotlib
seaborn
keras
tensorflow
pillow
opencv-python-headless
protobuf
dlib